Addition of the solutions to the second Andleab question for Day 2
The function reverseString will reverse any string that sent as its
argument.
It takes the length of the string and iterates through it once from the
last to the first.
Then it adds it to a new string. It returns true if the string is a
palindrome.

Test cases are also uploaded.