Addition of the solutions to the first Andleab question for Day 2
The function 'words' will split any the phrase into words and return the
number of times each word occured in the phrase. To split, it adds the
characters to a new string until it encounters a defined character (in
this case: whitespace, tab and new line)
Then to do the counts, it checks if the word already exists and if true,
it updates the value. Else, it adds the word as keys in an associative
array with the count as values.

Test cases are also uploaded.